+++ src/core/env.ts
@@ -36,7 +36,7 @@ class Env {
             // @ts-ignore
             this.dataSaver = window.navigator.connection.saveData;
             // @ts-ignore
-            navigator.connection.onchange = this.handleNetworkChange;
+            navigator.connection.onchange = this.handleNetworkChange.bind(this);
         }
 
         if ("deviceMemory" in navigator) {
@@ -48,6 +48,7 @@ class Env {
     private handleNetworkChange: EventListener = () => {
         // @ts-ignore
         this.connection = window.navigator.connection.effectiveType;
+        sessionStorage.removeItem("connection-choice");
     };
 
     /**
@@ -127,6 +128,38 @@ class Env {
     public setDOMState(newState: DOMState): void {
         this.domState = newState;
         document.documentElement.setAttribute("state", this.domState);
+    }
+
+    /**
+     * Checks if the provided connection is greater than or equal to the current conneciton.
+     * @param requiredConnection - network connection string
+     */
+    public checkConnection(requiredConnection): boolean {
+        let passed = false;
+        switch (requiredConnection) {
+            case "4g":
+                if (this.connection !== "2g" && this.connection !== "slow-2g" && this.connection !== "3g") {
+                    passed = true;
+                }
+                break;
+            case "3g":
+                if (this.connection !== "2g" && this.connection !== "slow-2g") {
+                    passed = true;
+                }
+                break;
+            case "2g":
+                if (this.connection !== "slow-2g") {
+                    passed = true;
+                }
+                break;
+            case "slow-2g":
+                passed = true;
+                break;
+            default:
+                passed = true;
+                break;
+        }
+        return passed;
     }
 
     private setBrowser() {